Adding support for a  "localsecrets" file for better auth
Added a "localsecrets" file for two reasons:
1) sometimes login fails if complex passwords are provided from command line
2) it is bad practice to have to supply a password as a parameter as it ends up in shell history

With these changes one can supply credentials (`user_email` and `user_password` in `localsecrets.py`), and if they are provided, the `--email` and `--password` parameters are initialised as optional. However if the parameters are supplied on command line it will override whatever is set in the `localsecrets.py` file.